Rename projectionReader
