+++ app/src/main/java/com/tranquilrock/androidscript/service/ClickService.kt
@@ -8,8 +8,6 @@ import android.graphics.Path
 import android.util.Log
 import android.view.accessibility.AccessibilityEvent
 import android.widget.Toast
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
@@ -19,63 +17,72 @@ class ClickService : AccessibilityService() {
         super.onServiceConnected()
         Log.d(TAG, "onServiceConnected")
 
-        instance = this
+        clicker = this
         Toast.makeText(this, "ClickService On", Toast.LENGTH_SHORT).show()
     }
 
-    class OffException : Exception("AccessibilityService Not On!")
+    override fun onDestroy() {
+        Log.d(TAG, "onDestroy")
+
+        clicker = null
+
+        Toast.makeText(this, "ClickService Down", Toast.LENGTH_SHORT).show()
+        stopService(Intent(this, WidgetService::class.java))
+        super.onDestroy()
+    }
+
+    suspend fun click(x: Int, y: Int) {
+        val path = Path().apply {
+            moveTo((x - 1).toFloat(), (y - 1).toFloat())
+            lineTo((x + 1).toFloat(), (y + 1).toFloat())
+        }
+        susDispatch(
+            GestureDescription.Builder()
+                .addStroke(StrokeDescription(path, 0, 1000))
+                .build()
+        )
+    }
+
+    suspend fun swipe(x1: Int, y1: Int, x2: Int, y2: Int) {
+        val path = Path().apply {
+            moveTo(x1.toFloat(), y1.toFloat())
+            lineTo(x2.toFloat(), y2.toFloat())
+        }
+        susDispatch(
+            GestureDescription.Builder()
+                .addStroke(StrokeDescription(path, 0, 1000))
+                .build()
+        )
+    }
+
+    private suspend fun susDispatch(gestureDescription: GestureDescription) {
+        clicker?.run {
+            suspendCoroutine<Unit?> { continuation ->
+                dispatchGesture(
+                    gestureDescription,
+                    object : GestureResultCallback() {
+                        override fun onCompleted(gestureDescription: GestureDescription?) =
+                            continuation.resume(null)
+
+                        override fun onCancelled(gestureDescription: GestureDescription?) {
+                            Log.i(TAG, "Cancelled: $gestureDescription")
+                            continuation.resume(null)
+                        }
+                    },
+                    null,
+                )
+            }
+        } ?: throw OffException()
+    }
 
     companion object {
         private val TAG = ClickService::class.java.simpleName
-        private var instance: ClickService? = null
 
-        val running: Boolean
-            get() = instance != null
-
-        suspend fun click(x: Int, y: Int) {
-            val path = Path().apply {
-                moveTo((x - 1).toFloat(), (y - 1).toFloat())
-                lineTo((x + 1).toFloat(), (y + 1).toFloat())
-            }
-            susDispatch(
-                GestureDescription.Builder()
-                    .addStroke(StrokeDescription(path, 0, 1000))
-                    .build()
-            )
-        }
-
-        suspend fun swipe(x1: Int, y1: Int, x2: Int, y2: Int) {
-            val path = Path().apply {
-                moveTo(x1.toFloat(), y1.toFloat())
-                lineTo(x2.toFloat(), y2.toFloat())
-            }
-            susDispatch(
-                GestureDescription.Builder()
-                    .addStroke(StrokeDescription(path, 0, 1000))
-                    .build()
-            )
-        }
-
-        private suspend fun susDispatch(gestureDescription: GestureDescription) {
-            instance?.run {
-                suspendCoroutine<Unit?> { continuation ->
-                    dispatchGesture(
-                        gestureDescription,
-                        object : GestureResultCallback() {
-                            override fun onCompleted(gestureDescription: GestureDescription?) =
-                                continuation.resume(null)
-
-                            override fun onCancelled(gestureDescription: GestureDescription?) {
-                                Log.i(TAG, "Cancelled: $gestureDescription")
-                                continuation.resume(null)
-                            }
-                        },
-                        null,
-                    )
-                }
-            } ?: throw OffException()
-        }
+        var clicker: ClickService? = null
+            private set
     }
+
+    class OffException : Exception("AccessibilityService Not On!")
 
     /**
      * AccessibilityEvent Callback, as all events got filtered, will not be called.
@@ -90,15 +97,5 @@ class ClickService : AccessibilityService() {
      * */
     override fun onInterrupt() {
         Log.d(TAG, "onInterrupt")
-    }
-
-    override fun onDestroy() {
-        Log.d(TAG, "onDestroy")
-
-        instance = null
-
-        Toast.makeText(this, "ClickService Down", Toast.LENGTH_SHORT).show()
-        stopService(Intent(this, WidgetService::class.java))
-        super.onDestroy()
     }
 }
